+++ honoka-json/honoka-json-api/src/main/java/de/honoka/sdk/json/api/service/JsonArrayService.java
@@ -2,9 +2,13 @@ package de.honoka.sdk.json.api.service;
 
 import de.honoka.sdk.json.api.JsonArray;
 
+import java.util.Collection;
+
 public interface JsonArrayService {
 
     <T> JsonArray<T> of(Class<T> clazz);
 
     <T> JsonArray<T> of(String jsonStr, Class<T> clazz);
+
+    <T> JsonArray<T> of(Collection<?> collection, Class<T> clazz);
 }