+++ honoka-json/honoka-json-api/src/main/java/de/honoka/sdk/json/api/service/JsonObjectService.java
@@ -7,4 +7,6 @@ public interface JsonObjectService {
     JsonObject of();
 
     JsonObject of(String jsonStr);
+
+    JsonObject of(Object obj);
 }