+++ honoka-json/honoka-json-fastjson/src/main/java/de/honoka/sdk/json/fastjson/FastJsonArrayService.java
@@ -1,7 +1,10 @@
 package de.honoka.sdk.json.fastjson;
 
+import com.alibaba.fastjson.JSON;
 import de.honoka.sdk.json.api.JsonArray;
 import de.honoka.sdk.json.api.service.JsonArrayService;
+
+import java.util.Collection;
 
 public class FastJsonArrayService implements JsonArrayService {
 
@@ -13,5 +16,10 @@ public class FastJsonArrayService implements JsonArrayService {
     @Override
     public <T> JsonArray<T> of(String jsonStr, Class<T> clazz) {
         return new FastJsonArray<>(jsonStr, clazz);
+    }
+
+    @Override
+    public <T> JsonArray<T> of(Collection<?> collection, Class<T> clazz) {
+        return of(JSON.toJSONString(collection), clazz);
     }
 }