+++ honoka-json/honoka-json-fastjson/src/main/java/de/honoka/sdk/json/fastjson/FastJsonObjectService.java
@@ -1,5 +1,6 @@
 package de.honoka.sdk.json.fastjson;
 
+import com.alibaba.fastjson.JSON;
 import de.honoka.sdk.json.api.JsonObject;
 import de.honoka.sdk.json.api.service.JsonObjectService;
 
@@ -13,5 +14,10 @@ public class FastJsonObjectService implements JsonObjectService {
     @Override
     public JsonObject of(String jsonStr) {
         return new FastJsonObject(jsonStr);
+    }
+
+    @Override
+    public JsonObject of(Object obj) {
+        return of(JSON.toJSONString(obj));
     }
 }