+++ honoka-json/honoka-json-gson/src/main/java/de/honoka/sdk/json/gson/GsonJsonArrayService.java
@@ -3,6 +3,8 @@ package de.honoka.sdk.json.gson;
 import de.honoka.sdk.json.api.JsonArray;
 import de.honoka.sdk.json.api.service.JsonArrayService;
 
+import java.util.Collection;
+
 public class GsonJsonArrayService implements JsonArrayService {
 
     @Override
@@ -13,5 +15,10 @@ public class GsonJsonArrayService implements JsonArrayService {
     @Override
     public <T> JsonArray<T> of(String jsonStr, Class<T> clazz) {
         return new GsonJsonArray<>(jsonStr, clazz);
+    }
+
+    @Override
+    public <T> JsonArray<T> of(Collection<?> collection, Class<T> clazz) {
+        return of(Common.gson.toJson(collection), clazz);
     }
 }